Отделяет слайдер эффекта от модуля выбора эффекта
