+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width,initial-scale=1" />
     <link rel="stylesheet" href="css/normalize.css" />
+    <link rel="stylesheet" href="nouislider/nouislider.css" />
     <link rel="stylesheet" href="css/style.css" />
     <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
     <title>Кекстаграм</title>
@@ -467,6 +468,7 @@
       </div>
     </template>
 
+    <script src="nouislider/nouislider.js"></script>
     <script src="js/main.js" type="module"></script>
   </body>
 </html>
